Added global defs for arrays to main.c
Also added these same defs as extern to AD9833.h   Best guess is that CMAKE version 3.22.1-1ubuntu1.22.04.1 is being strickter about extern vs. static interpretations of arrays.  It used to compile but 3.22.1 doesn't.  So I found it necessary to make the global defs in AD9833.h "extern" and have the true global def in main.c   hence the change here.